functions/create_sead/requirements.txt
appwrite==8.3.0
cryptography==41.0.0
python-dotenv==1.1.1
